feat(astro-kbve): diamond tier items + /api/mc-items + create-form picker (#11130)
Phase 5.3b of #10979. Fleshes out the Minecraft item database with the
full diamond tier (10 items) and wires the marketplace create form to
an autocomplete picker backed by the MDX collection.

Items (src/content/docs/mc/items/):
- diamond                — material; drop sources + storage form link
- diamond-block          — 9-diamond storage block (beacon base)
- diamond-pickaxe        — tier 5 mining; obsidian + ancient debris
- diamond-axe            — tier 5 mining + 9 damage melee
- diamond-shovel         — tier 5 dirt/sand/gravel/snow
- diamond-hoe            — tier 5 farming
- diamond-helmet         — 3 armor + 2 toughness
- diamond-chestplate     — 8 armor + 2 toughness
- diamond-leggings       — 6 armor + 2 toughness, supports swift_sneak
- diamond-boots          — 3 armor, supports depth/frost/soul_speed

Each MDX carries the full vanilla recipe inline (shaped grid layout),
compatible enchant list, durability, equipment slot/armor/toughness,
and mining tier where applicable.

API (src/pages/api/mc-items.json.ts):
- New `GET /api/mc-items.json` endpoint built at SSG time from the docs
  collection (filters `entry.id.startsWith('mc/items/')` + `data.mc_item`).
- Emits `{ items: [...], byRef: { ... }, count, generated_at }`. Sorted
  by display_name. Includes id, ref, slug, display_name, category,
  rarity, stack_size, tier, tags. Cache-Control: public, max-age=300.

UI (src/components/market/MCItemPicker.tsx):
- New `<MCItemPicker>` React component — autocomplete input with 24h
  localStorage cache of the items index.
- Substring match on `ref` + `display_name`. Shows display_name + ref +
  category in each dropdown row. Shows a `✓ <name>` deep-link badge
  when the typed ref resolves to a known item.
- Falls back to free-text mode if the seller types a ref we don't have
  yet (so they aren't gated on us writing the MDX first).

Create form (src/components/market/MarketCreateForm.tsx):
- Swaps the `Item ID` text input for `<MCItemPicker>` when `kind=mc_item`.
- Other kinds keep the raw text input.

Schema fix:
- MCDamage.attack_damage relaxed from int → float in both proto and Zod.
  MC tools deal fractional base damage (shovel +1.5, hoe +1.0) — int
  rejected diamond-shovel's 5.5 value at build time.

Theming: all picker styles use Starlight + KBVE brand vars
(`--sl-color-bg-sidebar`, `--sl-color-accent`, `--sl-color-hairline`).

+++ apps/kbve/astro-kbve/src/data/schema/mc/items.ts
@@ -26,9 +26,10 @@ export const MCEquipmentSchema = z.object({
 });
 export type MCEquipment = z.infer<typeof MCEquipmentSchema>;
 
-// proto: MCDamage
+// proto: MCDamage. MC tools deal fractional base damage (e.g. shovel +1.5)
+// — accept float; the in-game tooltip rounds to display.
 export const MCDamageSchema = z.object({
-	attack_damage: z.number().int().nonnegative(),
+	attack_damage: z.number().nonnegative(),
 	attack_speed: z.number(),
 });
 export type MCDamage = z.infer<typeof MCDamageSchema>;
